requirements.txt
azure.identity==1.15.0
azure.mgmt.cognitiveservices==13.5.0
azure-common==1.1.28
azure-core==1.29.5
azure-mgmt-core==1.4.0
